Moving contraptions in 1.15
- Ported the contraption renderer
- Fixed sticky blockstates of the linear chassis
- SuperByteBuffers can now be given a separate matrixStack for vertex lighting
- Kinetic TERs now make use of the passed light parameter

+++ src/main/java/com/simibubi/create/foundation/utility/AngleHelper.java
@@ -27,11 +27,11 @@ public class AngleHelper {
 		return facing == Direction.UP ? -90 : facing == Direction.DOWN ? 90 : 0;
 	}
 
-	public static float rad(float angle) {
+	public static float rad(double angle) {
 		return (float) (angle / 180 * Math.PI);
 	}
 
-	public static float deg(float angle) {
+	public static float deg(double angle) {
 		return (float) (angle * 180 / Math.PI);
 	}
